turn cache size into a 32-bit number

+++ include/SinglePoolCache.h
@@ -34,7 +34,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "CacheEntry.h"
 #include "ProcessorSerializer.h"
 
-template<template<auto, auto, auto, typename, bool> typename C, uint16_t numEntries, byte numAddressBits, byte numOffsetBits, typename T, bool debugMode = false>
+template<template<auto, auto, auto, typename, bool> typename C, uint32_t numEntries, byte numAddressBits, byte numOffsetBits, typename T, bool debugMode = false>
 class SinglePoolCache {
 private:
     using FakeCacheType = C<getNumberOfBitsForNumberOfEntries(numEntries), numAddressBits, numOffsetBits, T, debugMode>;
@@ -111,7 +111,7 @@ private:
  * @tparam T A static class that the cache data is read from and written to
  */
 template<template<auto, auto, auto, typename, bool> typename C,
-         uint16_t backingStoreSize,
+         uint32_t backingStoreSize,
          byte numAddressBits,
          byte numOffsetBits,
          typename T,
@@ -140,10 +140,10 @@ private:
     static inline UnderlyingCache_t theCache_;
 };
 
-template<template<auto, auto, auto, typename, bool> typename C, uint16_t backingStoreSize, byte numAddressBits, byte numOffsetBits, typename T, bool debugMode = false>
+template<template<auto, auto, auto, typename, bool> typename C, uint32_t backingStoreSize, byte numAddressBits, byte numOffsetBits, typename T, bool debugMode = false>
 using Cache_t = Cache<C, backingStoreSize, numAddressBits, numOffsetBits, T, debugMode>;
 
-template<template<auto, auto, auto, typename, bool> typename C, uint16_t backingStoreSize, byte numAddressBits, byte numOffsetBits, typename T, bool debugMode = false>
+template<template<auto, auto, auto, typename, bool> typename C, uint32_t backingStoreSize, byte numAddressBits, byte numOffsetBits, typename T, bool debugMode = false>
 using CacheInstance_t = typename Cache_t<C, backingStoreSize, numAddressBits, numOffsetBits, T, debugMode>::UnderlyingCache_t;
 
 #endif //SXCHIPSET_SINGLEPOOLCACHE_H
